+++ pymodbus/framer/base.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 from abc import abstractmethod
 
+from pymodbus.exceptions import ModbusIOException
 from pymodbus.factory import ClientDecoder, ServerDecoder
 from pymodbus.logging import Log
 from pymodbus.pdu import ModbusRequest, ModbusResponse
@@ -73,3 +74,45 @@ class FramerBase:
         data = message.function_code.to_bytes(1,'big') + message.encode()
         packet = self.encode(data, message.slave_id, message.transaction_id)
         return packet
+
+    def processIncomingPacket(self, data: bytes, callback, slave, tid=None):
+        """Process new packet pattern.
+
+        This takes in a new request packet, adds it to the current
+        packet stream, and performs framing on it. That is, checks
+        for complete messages, and once found, will process all that
+        exist.  This handles the case when we read N + 1 or 1 // N
+        messages at a time instead of 1.
+
+        The processed and decoded messages are pushed to the callback
+        function to process and send.
+        """
+        Log.debug("Processing: {}", data, ":hex")
+        self.databuffer += data
+        if self.databuffer == b'':
+            return
+        if not isinstance(slave, (list, tuple)):
+            slave = [slave]
+        while True:
+            if self.databuffer == b'':
+                return
+            used_len, data = self.decode(self.databuffer)
+            if used_len:
+                self.databuffer = self.databuffer[used_len:]
+            if not data:
+                return
+            if slave and 0 not in slave and self.incoming_dev_id not in slave:
+                Log.debug("Not a valid slave id - {}, ignoring!!", self.incoming_dev_id)
+                self.databuffer = b''
+                continue
+            if (result := self.decoder.decode(data)) is None:
+                self.databuffer = b''
+                raise ModbusIOException("Unable to decode request")
+            result.slave_id = self.incoming_dev_id
+            result.transaction_id = self.incoming_tid
+            Log.debug("Frame advanced, resetting header!!")
+            self.databuffer = self.databuffer[used_len:]
+            if tid and result.transaction_id and tid != result.transaction_id:
+                self.databuffer = b''
+            else:
+                callback(result)  # defer or push to a thread?
